Move compression header and response validation for transports into connect core
