+++ packages/connect-core/src/protocol-connect/index.ts
@@ -13,7 +13,10 @@
 // limitations under the License.
 
 export { codeFromHttpStatus, codeToHttpStatus } from "./http-status.js";
-export { createRequestHeader } from "./create-request-header.js";
+export {
+  createRequestHeader,
+  createRequestHeaderWithCompression,
+} from "./create-request-header.js";
 export {
   endStreamToJson,
   endStreamFromJson,
@@ -24,6 +27,9 @@ export { errorToJson } from "./error-to-json.js";
 export { parseContentType } from "./parse-content-type.js";
 export { parseTimeout } from "./parse-timeout.js";
 export { trailerDemux } from "./trailer-demux.js";
-export { validateResponse } from "./validate-response.js";
+export {
+  validateResponse,
+  validateResponseWithCompression,
+} from "./validate-response.js";
 export * from "./headers.js";
 export { protocolVersion } from "./version.js";
