+++ packages/connect-core/src/protocol-connect/validate-response.ts
@@ -17,6 +17,7 @@ import { Code } from "../code.js";
 import { parseContentType } from "./parse-content-type.js";
 import { codeFromHttpStatus } from "./http-status.js";
 import { ConnectError } from "../connect-error.js";
+import type { Compression } from "../compression.js";
 
 /**
  * Validates response status and header for the Connect protocol.
@@ -55,4 +56,32 @@ export function validateResponse(
     );
   }
   return { isConnectUnaryError: false };
+}
+
+export function validateResponseWithCompression(
+  methodKind: MethodKind,
+  useBinaryFormat: boolean,
+  acceptCompression: Compression[],
+  status: number,
+  headers: Headers
+): { compression: Compression | undefined; isConnectUnaryError: boolean } {
+  let compression: Compression | undefined;
+  const encodingField =
+    methodKind == MethodKind.Unary
+      ? "Content-Encoding"
+      : "Connect-Content-Encoding";
+  const encoding = headers.get(encodingField);
+  if (encoding != null && encoding.toLowerCase() !== "identity") {
+    compression = acceptCompression.find((c) => c.name === encoding);
+    if (!compression) {
+      throw new ConnectError(
+        `unsupported response encoding "${encoding}"`,
+        Code.InvalidArgument
+      );
+    }
+  }
+  return {
+    compression,
+    ...validateResponse(methodKind, useBinaryFormat, status, headers),
+  };
 }
