+++ packages/connect-core/src/protocol-grpc/validate-response.ts
@@ -17,6 +17,7 @@ import { parseContentType } from "./parse-content-type.js";
 import { ConnectError } from "../connect-error.js";
 import { findTrailerError } from "./trailer-status.js";
 import { Code } from "../code.js";
+import type { Compression } from "../compression.js";
 
 /**
  * Validates response status and header for the gRPC protocol.
@@ -48,4 +49,29 @@ export function validateResponse(
   if (err) {
     throw err;
   }
+}
+
+export function validateResponseWithCompression(
+  useBinaryFormat: boolean,
+  acceptCompression: Compression[],
+  status: number,
+  headers: Headers
+): { compression: Compression | undefined } {
+  validateResponse(useBinaryFormat, status, headers);
+
+  let compression: Compression | undefined;
+  const encodingField = "Grpc-Encoding";
+  const encoding = headers.get(encodingField);
+  if (encoding !== null && encoding.toLowerCase() !== "identity") {
+    compression = acceptCompression.find((c) => c.name === encoding);
+    if (!compression) {
+      throw new ConnectError(
+        `unsupported response encoding "${encoding}"`,
+        Code.InvalidArgument
+      );
+    }
+  }
+  return {
+    compression,
+  };
 }
